fixed employee main router
the problem here is that employee component was supposed to be the main
component for Employee Route. the problem is that employee component
has a nav but everytime router is directed to other child routes inside
employee, it completely replaces the employee component with another
child component. the fix is to set the other components (dashboard,
list) as children components inside the employeecomponent route